refactor: reformat code

+++ Hackavote.sol
@@ -12,16 +12,24 @@ contract Hackavote {
         opinions[msg.sender].push(opinion);
     }
 
-    function getOpinionOfHackerAtIndex(address hackerAddress, uint index) public view returns (bytes memory) {
+    function getOpinionOfHackerAtIndex(address hackerAddress, uint256 index)
+    public
+    view
+    returns (bytes memory)
+    {
         require(index < opinions[hackerAddress].length, "Index out of bounds.");
         return opinions[hackerAddress][index];
     }
 
-    function getOpinionCountOfHacker(address hackerAddress) public view returns (uint) {
+    function getOpinionCountOfHacker(address hackerAddress)
+    public
+    view
+    returns (uint256)
+    {
         return opinions[hackerAddress].length;
     }
 
-    function getHackerCount() public view returns (uint) {
+    function getHackerCount() public view returns (uint256) {
         return hackers.length;
     }
 }
